requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
pandas
yfinance
requests
google-genai
google-generativeai
pyodbc
openpyxl
feedparser
pyyaml
